Move Memoizer into Appear::Util
